refactor(simulator): reorganize rendering geometry primitives
Refactor the geometry rendering layer to make primitive drawing
more consistent and easier to maintain. This keeps rendering
responsibilities centralized in `Geometry.render` and preserves
existing behavior.
