+++ simulator/src/main/kotlin/life/sim/simulator/rendering/geometry/Line.kt
@@ -1,0 +1,10 @@
+package life.sim.simulator.rendering.geometry
+
+import com.badlogic.gdx.math.Vector2
+
+internal data class Line(
+    val a: Vector2,
+    val b: Vector2,
+    val width: Float,
+)
+
