+++ simulator/src/main/kotlin/life/sim/simulator/rendering/geometry/Rect.kt
@@ -1,0 +1,9 @@
+package life.sim.simulator.rendering.geometry
+
+internal data class Rect(
+    val x: Float,
+    val y: Float,
+    val width: Float,
+    val height: Float,
+)
+
