+++ simulator/src/main/kotlin/life/sim/simulator/rendering/geometry/Triangle.kt
@@ -1,0 +1,11 @@
+package life.sim.simulator.rendering.geometry
+
+internal data class Triangle(
+    val x1: Float,
+    val y1: Float,
+    val x2: Float,
+    val y2: Float,
+    val x3: Float,
+    val y3: Float,
+)
+
